+++ routes/settings.rb
@@ -48,6 +48,9 @@ get '/settings' do
   end
 
   @auth_types = %w(None Plain Login cram_md5)
+
+  # get hcbinpath (stored in config file vs db)
+  @hc_binpath = JSON.parse(File.read('config/agent_config.json'))['hc_binary_path']
  
   haml :global_settings
 end
@@ -58,19 +61,10 @@ post '/settings' do
     # Declare our db object first so that we can save values along the way instead of at the end
     hc_settings = HashcatSettings.first
 
-    # Hashcat Binary Path Sanity checks
-    if params[:hc_binpath].nil? || params[:hc_binpath].empty?
-      flash[:error] = 'You must set the path for your hashcat binary.'
-      redirect('/settings')
-    end
-
     unless File.file?(params[:hc_binpath])
       flash[:error] = 'Invalid file / path for hashcat binary.'
       redirect('/settings')
     end
-
-    # hcbinpath looks good
-    hc_settings.hc_binpath = params[:hc_binpath]
 
     # Max Task Time Sanity checks
     if params[:max_task_time].nil? || params[:max_task_time].empty?
@@ -137,7 +131,7 @@ post '/settings' do
     hc_settings.gpu_temp_retain = params[:gpu_temp_retain].to_i
 
     # Save force settings
-    params[:hc_force] == 'on' ? hc_settings.hc_force = '1' : hc_settings.hc_force = '0'
+    params[:hc_force] == 'on' ? hc_settings.hc_force = 1 : hc_settings.hc_force = 0
 
     hc_settings.save
 
@@ -195,6 +189,12 @@ post '/settings' do
       end
     end
     redirect to('/settings')
+  end
+ 
+  # distributed settings
+  if params[:chunk_size]
+    settings.chunk_size = params[:chunk_size].to_i
+    settings.save
   end
 
   flash[:success] = 'Settings updated successfully.'